+++ uni_modules/uv-checkbox/components/uv-checkbox/props.js
@@ -47,7 +47,7 @@ export default {
 		},
 		// label提示文字，因为nvue下，直接slot进来的文字，由于特殊的结构，无法修改样式
 		label: {
-			type: [String, Number],
+			type: [String, Number, Boolean],
 			default: ''
 		},
 		// label的字体大小，px单位