+++ uni_modules/uv-radio/components/uv-radio/props.js
@@ -42,7 +42,7 @@ export default {
 		},
 		// label提示文字，因为nvue下，直接slot进来的文字，由于特殊的结构，无法修改样式
 		label: {
-			type: [String, Number],
+			type: [String, Number, Boolean],
 			default: ''
 		},
 		// 整体的大小